rag_basic/Notebook/multimodal-rag-search/backend/services/requirements_oenai_pinecone.txt
# backend/requirements_openai_pinecone.txt
# Multimodal RAG with OpenAI + Pinecone + LangGraph
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# Pinecone
pinecone-client==3.0.0

# LangChain + LangGraph
langchain==0.1.0
langchain-core==0.1.14
langchain-community==0.0.13
langchain-openai==0.0.5  # OpenAI integration
langgraph==0.0.20
langgraph-checkpoint==0.0.1
langsmith==0.0.77

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Audio Processing
# Note: Whisper is via OpenAI API, no local install needed

# Video Processing
moviepy==1.0.3
opencv-python==4.9.0.80

# ML/AI Libraries (optional, for local processing)
numpy==1.26.3
torch==2.1.2  # Optional: for local models

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25

# Cache
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0